+++ src/membus/write.nim
@@ -5,12 +5,12 @@ from membus import TransactionSize, ENDIAN, Membus
 from memory_map import writeByte
 
  
-proc writeUint8*(membus : Membus, byte_address : uint32, value : uint8, endianness : ENDIAN) =
+proc writeUint8*(membus : Membus, byte_address : uint64, value : uint8, endianness : ENDIAN) =
   membus.writeByte(byte_address, value)
 
-proc writeUint16*(membus : Membus, byte_address : uint32, value : uint16, endianness : ENDIAN) =
+proc writeUint16*(membus : Membus, byte_address : uint64, value : uint16, endianness : ENDIAN) =
   var start_address = byte_address
-  var end_address = start_address + TransactionSize.HALFWORD.uint32 - 1
+  var end_address = start_address + TransactionSize.HALFWORD.uint64 - 1
 
   if endianness == ENDIAN.LITTLE:
     var byte_index = 0
@@ -26,9 +26,9 @@ proc writeUint16*(membus : Membus, byte_address : uint32, value : uint16, endian
       membus.writeByte(byte_address, value.bitsliced(b_slice).uint8)
       byte_index = byte_index + 1
 
-proc writeUint32*(membus : Membus, byte_address : uint32, value : uint32, endianness : ENDIAN) =
+proc writeUint32*(membus : Membus, byte_address : uint64, value : uint32, endianness : ENDIAN) =
   var start_address = byte_address
-  var end_address = start_address + TransactionSize.WORD.uint32 - 1
+  var end_address = start_address + TransactionSize.WORD.uint64 - 1
 
   if endianness == ENDIAN.LITTLE:
     var byte_index = 0
@@ -44,9 +44,9 @@ proc writeUint32*(membus : Membus, byte_address : uint32, value : uint32, endian
       membus.writeByte(byte_address, value.bitsliced(b_slice).uint8)
       byte_index = byte_index + 1
 
-proc writeUint64*(membus : Membus, byte_address : uint32, value : uint64, endianness : ENDIAN) =
+proc writeUint64*(membus : Membus, byte_address : uint64, value : uint64, endianness : ENDIAN) =
   var start_address = byte_address
-  var end_address = start_address + TransactionSize.DOUBLEWORD.uint32 - 1
+  var end_address = start_address + TransactionSize.DOUBLEWORD.uint64 - 1
 
   if endianness == ENDIAN.LITTLE:
     var byte_index = 0
